requirements.txt
asgiref==3.8.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
pillow==10.4.0
six==1.16.0
sqlparse==0.5.1
